+++ homelab_operator/webui/views.py
@@ -19,7 +19,7 @@ from .views_exp.service import create_service, edit_service, delete_service
 from .views_exp.network import create_network, edit_network, delete_network
 from .views_exp.schedule import create_schedule, edit_schedule, delete_schedule
 from .views_exp.shutdown_url import create_shutdown_url, edit_shutdown_url, delete_shutdown_url
-from .views_exp.wiki import create_wiki, edit_wiki, delete_wiki
+from .views_exp.wiki import public_wiki, create_wiki, edit_wiki, delete_wiki
 from .views_exp.uptime_statistic import create_uptime_statistic, delete_uptime_statistic, \
     reset_uptime_statistic